feat: add collectible moneys
