+++ src/components/Game/GameBoard/GameBoard.module.css
@@ -5,5 +5,4 @@
     height: 650px;
     background: #2a2c41;
     border: 2px solid #444;
-    overflow: hidden;
-}
+}
