+++ src/components/Game/Money/Money.jsx
@@ -1,6 +1,15 @@
+import styles from "./Money.module.css";
 
-const Money = () => {
-  return <div className="money">🪙</div>;
+const Money = ({ x, y }) => {
+  return (
+    <div
+      className={styles.money}
+      style={{
+        left: x + "px",
+        top: y + "px",
+      }}
+    />
+  );
 };
 
 export default Money;
